Use 127.0.0.1 instead of localhost
It is common that localhost also resolves to ::1 and it looks like
EventMachine is only trying the first address it gets back when
resolving the name: https://github.com/eventmachine/eventmachine/issues/885

RabbitMQ doesn't seem to listen on ::1 for me locally. This makes it
possible to start the app just by "foreman start".